+++ src/infrastructure/datasources/index.ts
@@ -1,1 +1,2 @@
 export * from "./file-log.datasource.impl.js";
+export * from "./mongo-log.datasource.impl.js";