+++ src/infrastructure/repositories/index.ts
@@ -1,1 +1,2 @@
 export * from "./file-log.repository.impl.js";
+export * from "./mongo-log.repository.impl.js";