+++ src/infrastructure/repositories/mongo-log.repository.impl.ts
@@ -1,0 +1,21 @@
+import type { FilterLogsOptions, LogDatasource, LogEntity, LogRepository } from "../../domain/index.js";
+
+export class MongoLogRepositoryImpl implements LogRepository {
+
+    constructor(
+        private readonly logDatasource: LogDatasource,
+    ) { };
+
+    readLogs(options?: FilterLogsOptions): Promise<LogEntity[]> {
+        return this.logDatasource.readLogs(options);
+    };
+
+    saveLog(log: LogEntity): Promise<void> {
+        return this.logDatasource.saveLog(log);
+    };
+
+    deleteLogs(options: FilterLogsOptions): Promise<void> {
+        return this.logDatasource.deleteLogs(options);
+    };
+
+};