+++ src/presentation/createLogger.ts
@@ -8,18 +8,27 @@ export const createLogger = async (options: CreateLoggerOptions = {}): Promise<L
 };
 
 export const buildTransport = async (options?: TransportOptions): Promise<LogRepository> => {
-    const type = options?.type ?? 'file';
 
-    switch (type) {
+    const resolvedOptions = options ?? { type: 'file' };
+
+    switch (resolvedOptions.type) {
         case 'file': {
             const { FileLogDatasourceImpl } = await import("../infrastructure/datasources/index.js");
             const { FileLogRepositoryImpl } = await import("../infrastructure/repositories/index.js");
 
-            const datasource = await FileLogDatasourceImpl.create(options);
+            const datasource = await FileLogDatasourceImpl.create(resolvedOptions);
             return new FileLogRepositoryImpl(datasource);
         };
+        case 'mongo': {
+            const { MongoLogDatasourceImpl } = await import("../infrastructure/datasources/index.js");
+            const { MongoLogRepositoryImpl } = await import("../infrastructure/repositories/index.js");
+
+            const datasource = await MongoLogDatasourceImpl.create(resolvedOptions);
+            return new MongoLogRepositoryImpl(datasource);
+        };
+        
         default: {
-            throw new Error(`Transport type "${type}" is not supported.`);
+            throw new Error(`Transport is not supported.`);
         };
     };
 };